+++ app/Http/Controllers/ItemPartialTransferManage.php
@@ -3,22 +3,14 @@
 namespace App\Http\Controllers;
 
 use App\Models\ItemPartialTransfer;
-use App\Models\Resource;
 use App\Models\Transformers\ItemPartialTransfer as ItemPartialTransferTransformer;
-use App\Option\Delete;
-use App\Option\Get;
-use App\Option\Post;
 use App\Response\Cache;
-use App\Response\Header\Headers;
-use App\Request\Parameter;
 use App\Request\Route;
-use App\Response\Pagination as UtilityPagination;
 use App\Request\Validate\ItemPartialTransfer as ItemPartialTransferValidator;
 use Exception;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Config;
 
 /**
  * Partial transfer of items
