+++ app/Http/Controllers/ItemTypeView.php
@@ -20,7 +20,7 @@ use Illuminate\Support\Facades\Config;
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ItemTypeController extends Controller
+class ItemTypeView extends Controller
 {
     protected bool $allow_entire_collection = true;
